♻️ Refactor: photolog 네비게이션을 feature 모듈별로 분리
