+++ app/build.gradle.kts
@@ -60,7 +60,9 @@ dependencies {
     implementation(projects.domain)
     implementation(projects.feature.login)
     implementation(projects.feature.main)
+    implementation(projects.feature.photolog.capture)
     implementation(projects.feature.photolog.detail)
+    implementation(projects.feature.photolog.editor)
     implementation(projects.feature.onboarding)
     implementation(projects.feature.goalEditor)
     implementation(projects.feature.goalManage)
