+++ core/util/src/main/java/com/twix/util/di/UtilModule.kt
@@ -1,9 +1,9 @@
 package com.twix.util.di
 
 import com.twix.util.bus.GoalRefreshBus
+import com.twix.util.bus.PhotologRefreshBus
 import com.twix.util.bus.StatsDetailRefreshBus
 import com.twix.util.bus.StatsRefreshBus
-import com.twix.util.bus.PhotologRefreshBus
 import org.koin.dsl.module
 
 val utilModule =
