+++ feature/photolog/detail/build.gradle.kts
@@ -5,8 +5,3 @@ plugins {
 android {
     namespace = "com.twix.photolog.detail"
 }
-
-dependencies {
-    implementation(projects.feature.photolog.capture)
-    implementation(projects.feature.photolog.editor)
-}
