+++ feature/photolog/detail/src/main/java/com/twix/photolog/detail/di/PhotologDetailModule.kt
@@ -2,17 +2,14 @@ package com.twix.photolog.detail.di
 
 import com.twix.navigation.NavRoutes
 import com.twix.navigation.base.NavGraphContributor
-import com.twix.photolog.capture.di.photologCaptureModule
 import com.twix.photolog.detail.PhotologDetailViewModel
-import com.twix.photolog.detail.navigation.PhotologGraph
-import com.twix.photolog.editor.di.photologEditorModule
+import com.twix.photolog.detail.navigation.PhotologDetailNavGraph
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 val photologDetailModule =
     module {
-        includes(photologCaptureModule, photologEditorModule)
         viewModelOf(::PhotologDetailViewModel)
-        single<NavGraphContributor>(named(NavRoutes.PhotologRoute.route)) { PhotologGraph }
+        single<NavGraphContributor>(named(NavRoutes.PhotologDetailRoute.route)) { PhotologDetailNavGraph }
     }
